Added in formatPlacesForApiCall
Added in this function to allow myself to formate the correct string for api calls to the distanceMatrix api by google. As the max destinations+origins can only equal to 25, to minimize api calls, I want to minimize amount of calls, so i split the matrix into 5 and 20, so total of 4 calls to populate a 20 by 20 matrix of all destinations.